reorganize based on review feedback
